refactor: update import paths from ra_aid.console.output to ra_aid.console.formatting for better organization and clarity fix: adjust cost formatting in get_cost_subtitle to display two decimal places for improved readability
