+++ ra_aid/tools/research.py
@@ -2,7 +2,7 @@ from langchain_core.tools import tool
 from rich.console import Console
 from rich.panel import Panel
 
-from ra_aid.console.output import console_panel
+from ra_aid.console.formatting import console_panel
 from ra_aid.database.repositories.trajectory_repository import get_trajectory_repository
 from ra_aid.database.repositories.human_input_repository import get_human_input_repository
